requirements/base.txt
fastapi==0.96.0
uvicorn==0.20.0
python-multipart==0.0.5
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
bcrypt==3.2.2
pydantic==2.1.1
jose==2.9.0
python-dotenv==1.0.0